requirements.txt
APScheduler==3.11.2
asgiref==3.11.1
boto3==1.42.73
botocore==1.42.73
cachetools==4.2.2
certifi==2026.2.25
charset-normalizer==3.4.6
Django==5.2.12
django-environ==0.13.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
jmespath==1.1.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2026.1.post1
redis==7.3.0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tornado==6.1
tzlocal==5.3.1
urllib3==2.6.3
